+++ routes/events/index.js
@@ -1,0 +1,4 @@
+// require('./create')
+module.exports = {
+  create: require('./create')
+ };
