+++ routes/index.js
@@ -1,0 +1,6 @@
+// require('./users');
+// require('./events');
+module.exports = {
+  users: require('./users'),
+  events: require('./events')
+ };
